Adds initial support for Git deployment on Heroku from git repo
Added Git functionality to locally clone a repository and push to a
remote URL. Heroku session is now able to clone a repository from (e.g.)
GitHub and push it to heroku repo corresponding to the application.

TODO: HerokuSession still have to pass the ITs.

TODO: This approach can be extended to do the deployment of an artifact.